# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GraphRAG Platform API
  description: Платформа мультимодального анализа корпоративных знаний (GraphRAG, закрытый контур). Контракт экспортируется
    в docs/api/openapi.yaml.
  version: 0.4.0
paths:
  /health:
    get:
      tags:
      - system
      summary: Проверка живости сервиса
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Health Health Get
  /metrics:
    get:
      tags:
      - system
      summary: Метрики Prometheus
      description: HTTP RPS/latency по маршрутам + бизнес-метрики чата/графа (этап 7, ADR-008).
      operationId: metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /auth/login:
    post:
      tags:
      - auth
      summary: 'Вход: выдаёт JWT (HS256) и создаёт сессию в PG'
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учётные данные или пользователь деактивирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/register:
    post:
      tags:
      - auth
      summary: 'Саморегистрация: роль viewer (только PUBLIC), JWT выдаётся сразу'
      description: 'Открытая регистрация умышленно даёт минимальную роль viewer.


        Расширение доступа (analyst/admin) — только через администратора:

        POST /admin/users. Аудит: auth.register.'
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '409':
          description: Имя пользователя занято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Неверный формат имени/пароля
  /auth/me:
    get:
      tags:
      - auth
      summary: Профиль текущего пользователя по Bearer-токену
      operationId: me_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Токен отсутствует, недействителен или истёк
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/ingest:
    post:
      tags:
      - admin
      summary: Запустить ingestion корпуса (background task)
      operationId: start_ingest_admin_ingest_post
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestStartResponse'
        '403':
          description: Не admin
        '409':
          description: Прогон уже идёт
  /admin/ingest/status:
    get:
      tags:
      - admin
      summary: Статус последнего/текущего прогона ingestion
      operationId: ingest_status_admin_ingest_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestStatusResponse'
        '403':
          description: Не admin
  /admin/stats:
    get:
      tags:
      - admin
      summary: 'Текущее наполнение хранилищ: Qdrant, Neo4j, PostgreSQL (агрегаты)'
      operationId: storage_stats_admin_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageStatsResponse'
        '403':
          description: Не admin
  /admin/users:
    get:
      tags:
      - admin
      summary: Список учётных записей (без секретов)
      operationId: list_users_admin_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '403':
          description: Не admin
    post:
      tags:
      - admin
      summary: Создать учётную запись с указанной ролью
      description: 'Роль задаёт метки доступа: viewer→PUBLIC, analyst→PUBLIC+INTERNAL, admin→все.'
      operationId: add_user_admin_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '403':
          description: Не admin
        '409':
          description: Имя занято
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/chat:
    post:
      tags:
      - chat
      summary: Вопрос агенту GraphRAG (SSE-стриминг или JSON при stream=false)
      description: 'Конвейер: guardrails -> planner -> retrieval (вектор+граф, ACL pre-fetch) -> fusion -> rerank -> generate
        -> evaluate (re-plan <= APP_AGENT_MAX_ITERATIONS) -> guardrail_out (проверка цитат).


        При stream=true возвращает text/event-stream: события `status`, `token`, `done`, `error` (формат - event:\ndata:\n\n).
        При stream=false - обычный JSON.'
      operationId: chat_api_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Chat Api Chat Post
        '401':
          description: Нет/невалидный JWT
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ChatRequest:
      properties:
        message:
          type: string
          maxLength: 8000
          minLength: 1
          title: Message
          description: Вопрос пользователя (санитайзер и guardrails применяются на сервере)
          examples:
          - Какие требования к раскрытию информации эмитентами?
        session_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Session Id
          description: Идентификатор диалога; None - начать новый диалог
        stream:
          type: boolean
          title: Stream
          description: true - SSE-поток событий, false - JSON
          default: true
      type: object
      required:
      - message
      title: ChatRequest
      description: Запрос к агенту; ``stream=False`` возвращает обычный JSON (Postman/нагрузка).
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          anyOf:
          - items:
              additionalProperties: true
              type: object
            type: array
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ErrorResponse
      description: Тело ошибки домена/валидации.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IngestStartResponse:
      properties:
        state:
          type: string
          const: started
          title: State
          default: started
        message:
          type: string
          title: Message
          default: Прогон ingestion запущен
      type: object
      title: IngestStartResponse
    IngestStatusResponse:
      properties:
        state:
          type: string
          enum:
          - idle
          - running
          - done
          - error
          title: State
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        stats:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Stats
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - state
      title: IngestStatusResponse
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[a-zA-Z0-9_.-]+$
          title: Username
          description: Имя пользователя
        password:
          type: string
          maxLength: 72
          minLength: 8
          title: Password
          description: Пароль (bcrypt на сервере)
      type: object
      required:
      - username
      - password
      title: LoginRequest
      description: Тело запроса логина.
      examples:
      - password: analyst123
        username: analyst
    MeResponse:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          title: Role
        clearances:
          items:
            type: string
          type: array
          title: Clearances
          description: Метки доступа роли (ACL для pre-fetch фильтров)
      type: object
      required:
      - user_id
      - username
      - role
      - clearances
      title: MeResponse
      description: Профиль текущего пользователя (по Bearer-токену).
    Neo4jStats:
      properties:
        acts:
          type: integer
          title: Acts
        authorities:
          type: integer
          title: Authorities
        topics:
          type: integer
          title: Topics
        concepts:
          type: integer
          title: Concepts
        relationships:
          type: integer
          title: Relationships
      type: object
      required:
      - acts
      - authorities
      - topics
      - concepts
      - relationships
      title: Neo4jStats
      description: Узлы и рёбра графа знаний (по меткам онтологии ADR-006).
    PostgresStats:
      properties:
        users:
          type: integer
          title: Users
        chat_sessions:
          type: integer
          title: Chat Sessions
        chat_messages:
          type: integer
          title: Chat Messages
      type: object
      required:
      - users
      - chat_sessions
      - chat_messages
      title: PostgresStats
      description: Учётные записи и история диалогов.
    QdrantStats:
      properties:
        collection:
          type: string
          title: Collection
        points:
          type: integer
          title: Points
      type: object
      required:
      - collection
      - points
      title: QdrantStats
      description: Наполнение векторного хранилища.
    StorageStatsResponse:
      properties:
        qdrant:
          $ref: '#/components/schemas/QdrantStats'
        neo4j:
          $ref: '#/components/schemas/Neo4jStats'
        postgres:
          $ref: '#/components/schemas/PostgresStats'
      type: object
      required:
      - qdrant
      - neo4j
      - postgres
      title: StorageStatsResponse
      description: Снимок текущего наполнения хранилищ (без прогресса прогонов).
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          const: bearer
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
          description: TTL токена, секунды
        username:
          type: string
          title: Username
        role:
          type: string
          title: Role
      type: object
      required:
      - access_token
      - expires_in
      - username
      - role
      title: TokenResponse
      description: 'Успешный логин: JWT + метаданные сессии.'
      examples:
      - access_token: eyJhbGciOiJIUzI1NiIs...
        expires_in: 1800
        role: analyst
        token_type: bearer
        username: analyst
    UserCreate:
      properties:
        username:
          type: string
          maxLength: 64
          minLength: 3
          pattern: ^[a-zA-Z0-9_.-]+$
          title: Username
          description: Имя пользователя (уникально)
        password:
          type: string
          maxLength: 72
          minLength: 8
          title: Password
          description: Пароль (bcrypt на сервере)
        role:
          type: string
          enum:
          - viewer
          - analyst
          - admin
          title: Role
          description: Роль определяет метки доступа
          default: viewer
      type: object
      required:
      - username
      - password
      title: UserCreate
      description: Создание учётной записи администратором.
      examples:
      - password: qa123456
        role: viewer
        username: qa_viewer
    UserOut:
      properties:
        user_id:
          type: string
          title: User Id
        username:
          type: string
          title: Username
        role:
          type: string
          enum:
          - viewer
          - analyst
          - admin
          title: Role
        clearances:
          items:
            type: string
          type: array
          title: Clearances
          description: Метки доступа роли
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - user_id
      - username
      - role
      - clearances
      - is_active
      - created_at
      title: UserOut
      description: Учётная запись без секретов.
    UsersListResponse:
      properties:
        users:
          items:
            $ref: '#/components/schemas/UserOut'
          type: array
          title: Users
      type: object
      required:
      - users
      title: UsersListResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
tags:
- name: system
  description: Здоровье сервиса и метрики
- name: auth
  description: Аутентификация и профиль пользователя
- name: admin
  description: Управление ingestion корпуса (роль admin)
- name: chat
  description: Вопросы к GraphRAG-агенту (SSE-стриминг)